+++ packages/browser/__tests__/Session.spec.ts
@@ -260,6 +260,38 @@ describe("Session", () => {
       );
     });
 
+    it("logs the session out when its tokens expire", async () => {
+      jest.useFakeTimers();
+      const MOCK_TIMESTAMP = 10000;
+      jest.spyOn(Date, "now").mockReturnValueOnce(MOCK_TIMESTAMP);
+      const clientAuthentication = mockClientAuthentication();
+      const incomingRedirectHandler = jest.spyOn(
+        clientAuthentication,
+        "handleIncomingRedirect"
+      );
+      incomingRedirectHandler.mockResolvedValueOnce({
+        isLoggedIn: true,
+        sessionId: "Arbitrary session ID",
+        expirationDate: MOCK_TIMESTAMP + 1337,
+      });
+      const mySession = new Session({ clientAuthentication });
+      await mySession.handleIncomingRedirect("https://some.url");
+      expect(setTimeout).toHaveBeenCalledWith(expect.any(Function), 1337);
+      expect(mySession.info.isLoggedIn).toBe(true);
+
+      // Usually, we'd be able to use `jest.runAllTimers()` here. However,
+      // logout happens asynchronously. While this is inconsequential in
+      // running apps (the timeout complets asynchronously anyway), here, we can
+      // take advantage that we return the Promise from the callback in
+      // `setTimeout`, so that we can `await` it in this test before checking
+      // whether logout was successful:
+      const expireTimeout = ((setTimeout as unknown) as jest.Mock<
+        typeof setTimeout
+      >).mock.calls[0][0];
+      await expireTimeout();
+      expect(mySession.info.isLoggedIn).toBe(false);
+    });
+
     it("leaves the session's info unchanged if no session is obtained after redirect", async () => {
       const clientAuthentication = mockClientAuthentication();
       clientAuthentication.handleIncomingRedirect = jest.fn(
@@ -271,29 +303,23 @@ describe("Session", () => {
       expect(mySession.info.sessionId).toEqual("mySession");
     });
 
-    function sleep(ms: number): Promise<void> {
-      return new Promise((resolve) => setTimeout(resolve, ms));
-    }
-
     it("prevents from hitting the token endpoint twice with the same auth code", async () => {
       const clientAuthentication = mockClientAuthentication();
       const obtainedSession: ISessionInfo = {
         isLoggedIn: true,
         sessionId: "mySession",
       };
-      let secondRequestIssued = false;
+      let continueAfterSecondRequest: (value?: unknown) => void;
+      const blockingResponse = new Promise((resolve, _reject) => {
+        continueAfterSecondRequest = resolve;
+      });
       const blockingRequest = async (): Promise<ISessionInfo> => {
-        while (!secondRequestIssued) {
-          // eslint-disable-next-line no-await-in-loop
-          await sleep(100);
-        }
+        await blockingResponse;
         return obtainedSession;
       };
       // The ClientAuthn's handleIncomingRedirect will only return when the
       // second Session's handleIncomingRedirect has been called.
-      clientAuthentication.handleIncomingRedirect = jest.fn(
-        async (_url: string) => blockingRequest()
-      );
+      clientAuthentication.handleIncomingRedirect = jest.fn(blockingRequest);
       const mySession = new Session({ clientAuthentication });
       const firstTokenRequest = mySession.handleIncomingRedirect(
         "https://my.app/?code=someCode&state=arizona"
@@ -301,7 +327,9 @@ describe("Session", () => {
       const secondTokenRequest = mySession.handleIncomingRedirect(
         "https://my.app/?code=someCode&state=arizona"
       );
-      secondRequestIssued = true;
+      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+      // We know that it has been set by the call to `blockingRequest`.
+      continueAfterSecondRequest!();
       const tokenRequests = await Promise.all([
         firstTokenRequest,
         secondTokenRequest,